Qc refactor

+++ genotools/__main__.py
@@ -6,6 +6,8 @@ import os
 import tempfile
 import pandas as pd
 import json
+import warnings
+import pathlib
 
 from genotools.utils import shell_do, upfront_check
 from genotools.qc import SampleQC, VariantQC
@@ -23,7 +25,9 @@ def execute_pipeline(steps, steps_dict, geno_path, out_path, samp_qc, var_qc, an
 
     # otherwise tmpdir
     else:
-        step_paths = [f'{tmp_dir.name}/out']
+        out_path_pathlib = pathlib.PurePath(out_path)
+        out_path_name = out_path_pathlib.name
+        step_paths = [f'{tmp_dir.name}/{out_path_name}']
 
     # if first step is ancestry, make new steps list to call within-ancestry
     if steps[0] == 'ancestry':
@@ -69,9 +73,17 @@ def execute_pipeline(steps, steps_dict, geno_path, out_path, samp_qc, var_qc, an
                 for geno, label in zip(out_dict[step]['output']['split_paths'], out_dict[step]['data']['labels_list']):
                     out_dict[label] = execute_pipeline(steps_ancestry, steps_dict, geno, f'{out_path}_{label}', samp_qc, var_qc, ancestry, assoc, args, tmp_dir)
 
-        # keep track of paths
         else:
-            step_paths.append(step_output)
+            # if warn is True and step input doesn't exist print error and reset step input
+            if args['warn'] and (not os.path.isfile(f'{step_input}.pgen')) and (len(step_paths) > 1):
+                warnings.warn(f'{step_input}.pgen was not created. Continuing to next step...')
+                step_input = f'{step_paths[-2]}' if step != steps[1] else geno_path
+                step_output = f'{step_paths[-2]}_{step}' if step != steps[-1] else out_path
+                step_paths.append(step_output)
+            
+            # otherwise keep track of paths
+            else:
+                step_paths.append(step_output)
 
         # samp qc setup and call
         if step in samp_steps:
@@ -114,21 +126,29 @@ def execute_pipeline(steps, steps_dict, geno_path, out_path, samp_qc, var_qc, an
         
         # remove old files when appropriate 
         if (not args['full_output']):
-            remove_step_index = step_paths.index(step_output) - 1
-            remove_path = step_paths[remove_step_index]
-            if os.path.isfile(f'{remove_path}.pgen'):
-                os.remove(f'{remove_path}.pgen')
-                os.remove(f'{remove_path}.psam')
-                os.remove(f'{remove_path}.pvar')
+            # when warn is True and step fails, don't remove old file
+            if args['warn'] and (not out_dict[step]['pass']):
+                remove = False
+            else:
+                remove = True
+                remove_step_index = step_paths.index(step_output) - 1
+
+            if remove:
+                remove_path = step_paths[remove_step_index]
+                # make sure we're not removing the output
+                if os.path.isfile(f'{remove_path}.pgen') and (remove_path != out_path):
+                    os.remove(f'{remove_path}.pgen')
+                    os.remove(f'{remove_path}.psam')
+                    os.remove(f'{remove_path}.pvar')
 
     out_dict['paths'] = step_paths
 
     return out_dict
 
 
-def build_metrics_pruned_df(metrics_df, pruned_df, dictionary, ancestry='all'):
+def build_metrics_pruned_df(metrics_df, pruned_df, gwas_df, dictionary, ancestry='all'):
     #TODO: Add association output
-    for step in ['callrate', 'sex', 'realted', 'het', 'case_control', 'haplotype', 'hwe', 'geno']:
+    for step in ['callrate', 'sex', 'realted', 'het', 'case_control', 'haplotype', 'hwe', 'geno','ld']:
         if step in dictionary.keys():
             qc_step = dictionary[step]['step']
             pf = dictionary[step]['pass']
@@ -146,10 +166,14 @@ def build_metrics_pruned_df(metrics_df, pruned_df, dictionary, ancestry='all'):
                 level = 'variant'
 
             for metric, value in dictionary[step]['metrics'].items():
-                tmp_metrics_df = pd.DataFrame({'step':[qc_step], 'pruned_count':[value], 'metric':[metric], 'ancestry':[ancestry_label], 'level':[level], 'pass': [pf]})
-                metrics_df = pd.concat([metrics_df, tmp_metrics_df], ignore_index=True)
+                if step == 'assoc':
+                    tmp_gwas_df = pd.DataFrame({'value':[value], 'metric':[metric], 'ancestry':[ancestry_label]})
+                    gwas_df = pd.concat([gwas_df, tmp_gwas_df], ignore_index=True)
+                else:
+                    tmp_metrics_df = pd.DataFrame({'step':[qc_step], 'pruned_count':[value], 'metric':[metric], 'ancestry':[ancestry_label], 'level':[level], 'pass': [pf]})
+                    metrics_df = pd.concat([metrics_df, tmp_metrics_df], ignore_index=True)
 
-    return metrics_df, pruned_df
+    return metrics_df, pruned_df, gwas_df
 
 
 if __name__=='__main__':
@@ -161,6 +185,7 @@ if __name__=='__main__':
     parser.add_argument('--out_path', type=str, nargs='?', default=None, const=None, help='Prefix for output (including path)', required=True)
     parser.add_argument('--full_output', type=str, nargs='?', default='True', const='True', help='Output everything')
     parser.add_argument('--skip_fails', type=str, nargs='?', default='False', const='True', help='Skip up front check for fails')
+    parser.add_argument('--warn', type=str, nargs='?', default='False', const='True', help='Warn of error and continue running pipeline')
 
     # ancerstry arguments
     parser.add_argument('--ancestry', type=str, nargs='?', default='False', const='True', help='Split by ancestry')
@@ -255,6 +280,7 @@ if __name__=='__main__':
         args_dict['haplotype'] = 1e-4
         args_dict['hwe'] = 1e-4
         args_dict['filter_controls'] = True
+        args_dict['ld'] = None
 
     print(args_dict)
 
@@ -297,6 +323,7 @@ if __name__=='__main__':
     clean_out_dict = dict()
     metrics_df = pd.DataFrame()
     pruned_df = pd.DataFrame()
+    gwas_df = pd.DataFrame()
 
     if 'ancestry' in out_dict.keys():
         ancestry_counts_df = pd.DataFrame(out_dict['ancestry']['metrics']['predicted_counts']).reset_index()
@@ -319,16 +346,17 @@ if __name__=='__main__':
     
         for ancestry in ['AFR', 'SAS', 'EAS', 'EUR', 'AMR', 'AJ', 'CAS', 'MDE', 'FIN', 'AAC']:
             if ancestry in out_dict.keys():
-                metrics_df, pruned_df = build_metrics_pruned_df(metrics_df=metrics_df, pruned_df=pruned_df, dictionary=out_dict[ancestry], ancestry=ancestry)
+                metrics_df, pruned_df, gwas_df = build_metrics_pruned_df(metrics_df=metrics_df, pruned_df=pruned_df, gwas_df=gwas_df, dictionary=out_dict[ancestry], ancestry=ancestry)
 
     else:
-        metrics_df, pruned_df = build_metrics_pruned_df(metrics_df=metrics_df, pruned_df=pruned_df, dictionary=out_dict)
+        metrics_df, pruned_df, gwas_df = build_metrics_pruned_df(metrics_df=metrics_df, pruned_df=pruned_df, gwas_df=gwas_df, dictionary=out_dict)
 
     # for weird error with the first sample in pruned file showing up twice when run in tmp file
     pruned_df = pruned_df.drop_duplicates(subset=['#FID','IID'], ignore_index=True)
 
     clean_out_dict['QC'] = metrics_df.to_dict()
     clean_out_dict['pruned_samples'] = pruned_df.to_dict()
+    clean_out_dict['GWAS'] = gwas_df.to_dict()
 
     # dump output to json
     with open(f'{args_dict["out_path"]}.json', 'w') as f:
